% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_formatted_worksheet.R
\name{add_formatted_worksheet}
\alias{add_formatted_worksheet}
\title{Add excel sheet with a formatted header}
\usage{
add_formatted_worksheet(
  data,
  workbook,
  sheet,
  wrapHeadlineText = FALSE,
  collabels = TRUE,
  colwidths = TRUE,
  standards = NULL,
  dbsource = deparse(substitute(data))
)
}
\arguments{
\item{data}{the Data frame to export to the Excel sheet}

\item{workbook}{The woorkbook object}

\item{sheet}{The Excel sheet name}

\item{wrapHeadlineText}{Should headline allow wrapping of text. TRUE / FALSE, standard is FALSE}

\item{collabels}{Should headline be changed to standard labels, TRUE / FALSE, standard is TRUE}

\item{colwidths}{Should defined standard colwidths be used. TRUE / FALSE, standard is TRUE}

\item{standards}{tables with column_standards}

\item{dbsource}{database source of data in column standards table}
}
\value{
None. A new sheet with formatted headline is added to the workbook object with
}
\description{
Add excel sheet with a header formatted with bold font, column labels and column width. The function is a wrapper for
\code{openxlsx::addWorksheet} and are using \code{NVIdb::standardize_columns} to standardize columns.
}
\details{
Add excel sheet with a formatted header. The header will be formatted with bold text.

\code{collabels = TRUE} the column names are translated to column labels in accord with the column standards table, see \code{standardize_columns}.

\code{colwidths = TRUE} the column widths are given in accord with the column standards table, see \code{standardize_columns}.

\code{wrapHeadlineText = TRUE} theheadline text is allowed to wrap on two or more lines. The parameter should be chosen in accord with what looks
nice depending on column labels and column widths.

\code{standards} is the name of the table with column standards. If no parameter is given, the columns_standards.csv is used.  column names are translated to column labels in accord with the column standards table, see \code{standardize_columns}.

\code{dbsource} is the dbsource in the column standards table making it possible to tailer the column laels and column widths per table.
}
\examples{
\dontrun{
#

}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
