% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_text_line.R
\name{append_text_line}
\alias{append_text_line}
\title{Append new last row with text}
\usage{
append_text_line(data, text, empty_rows = 0)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
The data that should get a new row with the text.}

\item{text}{[\code{character(1)}]\cr
The text that should be appended.}

\item{empty_rows}{[\code{integer(1)}]\cr
Number of empty rows between the data and the text. Defaults to 0.}
}
\value{
A data frame with two more rows, one empty and one with generated data in the first column.
}
\description{
Function to add new row with text in a data frame.
    This is intended for appending explanation text (footnotes etc.)
    to data frames that should be published in Excel.
}
\details{
Two rows are appended to the data frame, the first is empty, the second has the generated date in the first column. The
}
\examples{
\dontrun{
# Add row with generated date using standard values
gris_virus_slaktegris_utvalg <- append_date_generated_line(gris_virus_slaktegris_utvalg)
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
