% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIpretty-package.R
\docType{package}
\name{NVIpretty-package}
\alias{NVIpretty}
\alias{NVIpretty-package}
\title{NVIpretty: Tools for making R-output pretty in accord with NVI's graphical profile}
\description{
Provides tools for styling output from R in accord with NVI's graphical profile. Included are colour palettes and a NVI-theme for ggplot2 as well as tools to format and style output in Excel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NorwegianVeterinaryInstitute/NVIpretty}
  \item Report bugs at \url{https://github.com/NorwegianVeterinaryInstitute/NVIpretty/issues}
}

}
\author{
\strong{Maintainer}: Petter Hopp \email{Petter.Hopp@vetinst.no}

Other contributors:
\itemize{
  \item Norwegian Veterinary Institute [copyright holder]
}

}
\keyword{internal}
