% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_gradient_green}
\alias{palette_gradient_green}
\alias{palette_gradient_red}
\alias{palette_gradient_air}
\alias{palette_gradient_darkblue}
\alias{palette_gradient_earth}
\alias{palette_gradient_food}
\alias{palette_gradient_graa}
\alias{palette_gradient_grey}
\alias{palette_gradient_gul}
\alias{palette_gradient_jord}
\alias{palette_gradient_lightblue}
\alias{palette_gradient_luft}
\alias{palette_gradient_mat}
\alias{palette_gradient_vann}
\alias{palette_gradient_water}
\alias{palette_gradient_yellow}
\title{Palettes with colour gradients for NVI for 2025 - today's date}
\format{
A vector with unnamed colours:
\describe{
  \item{palette_gradient_green}{Green colour gradient based on the NVI colour "Jord".}
  \item{palette_gradient_jord}{Alternative name for the green colour gradient based on the NVI colour "Jord".}
  \item{palette_gradient_earth}{Alternative name for the green colour gradient based on the NVI colour "Jord".}
  \item{palette_gradient_darkblue}{Blue colour gradient based on the NVI colour "Vann".}
  \item{palette_gradient_vann}{Alternative name for the blue colour gradient based on the NVI colour "Vann".}
  \item{palette_gradient_water}{Alternative name for the blue colour gradient based on the NVI colour "Vann".}
  \item{palette_gradient_lightblue}{Blue colour gradient based on the NVI colour "Luft".}
  \item{palette_gradient_luft}{Alternative name for the blue colour gradient based on the NVI colour "Luft".}
  \item{palette_gradient_air}{Alternative name for the blue colour gradient based on the NVI colour "Luft".}
  \item{palette_gradient_red}{Red colour gradient based on the NVI colour "Mat".}
  \item{palette_gradient_mat}{Alternative name for the red colour gradient based on the NVI colour "Mat".}
  \item{palette_gradient_food}{Alternative name for the red colour gradient based on the NVI colour "Mat".}
  \item{palette_gradient_yellow}{Yellow colour gradient based on the NVI colour "Gul".}
  \item{palette_gradient_gul}{Alternative name for the yellow colour gradient based on the NVI colour "Gul".}
  \item{palette_gradient_grey}{Grey colour gradient based on the NVI colour "Svart".}
  \item{palette_gradient_graa}{Alternative name for the grey colour gradient based on the NVI colour "Svart".}
}

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes.R".

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.
}
\source{
Grafisk profil for Veterin<U+00E6>rinstituttet version 2024 (available on
  NVI's intranet).
}
\usage{
palette_gradient_green

palette_gradient_red

palette_gradient_air

palette_gradient_darkblue

palette_gradient_earth

palette_gradient_food

palette_gradient_graa

palette_gradient_grey

palette_gradient_gul

palette_gradient_jord

palette_gradient_lightblue

palette_gradient_luft

palette_gradient_mat

palette_gradient_red

palette_gradient_vann

palette_gradient_water

palette_gradient_yellow
}
\description{
Colour gradients with six levels for six different base colours.
    The colours in the NVI profile are used as base colours and the gradients
    are developed from these.
}
\details{
The colour gradients are intended for choropleth maps, graphs and
    other figures where different gradients of one colour is wanted.

There are six different gradients, each with six colours. Each palette is
    given a name of the form "palette_gradient_xxxxxx" where xxxxxx is the
    name of the base colour. To make it easier to remember the gradient name,
    each gradient is represented by several palettes, for example the green
    gradient based on the NVI colour "Jord" is represented by the palettes:
    palette_gradient_jord, palette_gradient_earth, and palette_gradient_green.

The colours are displayed in the vignette "NVI colour palettes".
}
\keyword{datasets}
