% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile_2020.R
\docType{data}
\name{palette_graph_2020}
\alias{palette_graph_2020}
\title{Old palette with colours for graphs for 2020-2024}
\format{
A vector with named colours:
\describe{
  \item{colours}{13 different colours}
}
}
\source{
Grafisk profil for Veterinærinstituttet versjon 2020.
}
\usage{
palette_graph_2020
}
\description{
A vector with the hex values for colours to be used in graphs.
}
\details{
The colour codes are based on the hex values in the Graphical profile
    handbook except for the body text where the cmyk value is used. There is
    not agreement between the hex and cmyk values, but hex seem to give the
    most correct colour.

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes_2020.R".
}
\keyword{datasets}
