% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_NVI}
\alias{palette_NVI}
\title{Palette with NVI's profile colours for 2025 - today's date}
\format{
A vector with named colours:
\describe{
\item{#59CD8B: Jord}{The green colour that signifies NVI's core area "Terestrial animals".}
\item{#D7F4FF: Luft}{A light blue colour".}
\item{#1C4FB9: Vann}{The blue colour that signifies NVI's core area "Aquatic animals".}
\item{#F7FDFF: Hvit}{Whitish. It is used in headings with a coloured or black background.}
\item{#091A3E: Svart}{Blackish for use in body text in reports etc. It can also be used in headings.}
\item{#BCEED1: Lys jord}{Used in graphs.}
\item{#95D9F3: M&#U00F8rk luft}{Used in graphs.}
\item{#C7D9FF: Lyst vann}{Used in graphs.}
\item{#FF5447: Mat}{The red colour that signifies NVI's core area "Food and feed safety".}
\item{#FFDAD4: Lys mat}{Used in graphs.}
}
}
\source{
Grafisk profil for Veterin<U+00E6>rinstituttet version 2024 (available on
NVI's intranet).
}
\usage{
palette_NVI
}
\description{
A vector with the hex values for NVI's profile colours.
}
\details{
The colour codes are based on the hex values in the official colour profile.

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes.R".
}
\keyword{datasets}
