% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_NVI_profile}
\alias{palette_NVI_profile}
\title{Palette with NVI's profile colours}
\format{
A vector with named colours:
\describe{
\item{blue}{The blue colour that signifies NVI's core area "Aquatic animals".}
\item{green}{The green colour that signifies NVI's core area "Terestrial animals".}
\item{red}{The red colour that signifies NVI's core area "Food and feed safety".}
\item{bodytext}{Black 80\% for use in body text in reports etc. It can also be used in headings.}
\item{headingblue}{Blue alternative for use in headings.}
\item{background}{Background colour for large areas as on stands.}
}
}
\source{
Grafisk profil for Veterinærinstituttet versjon 2020.
}
\usage{
palette_NVI_profile
}
\description{
A vector with the hex values for NVI's profile colours.
}
\details{
The colour codes are based on the hex values except for the body text where the cmyk value is used.
There is not agrement between the hex and cmyk values, but hex seem to give the most correct colour.

\if{html}{\out{<div class="sourceCode">}}\preformatted{The script for preparing the palette is given in \code{./data-raw/generate_NVI_palettes.R}.
}\if{html}{\out{</div>}}
}
\keyword{datasets}
