% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_table_blue}
\alias{palette_table_blue}
\title{Palette with colours for table with blue background}
\format{
A vector with named colours:
\describe{
\item{blue1}{Background colour for heading.}
\item{blue2}{Background colour for sub-heading.}
\item{blue3}{Background colour for sub-heading.}
\item{blue4}{Background colour for table text.}
\item{blue5}{Preferred background colour for table text.}
\item{bodytext}{Black 70\% font colour for use in body text.}
\item{heading}{White font colour for headings.}
}
}
\source{
Grafisk profil for Veterinærinstituttet versjon 2020.
}
\usage{
palette_table_blue
}
\description{
A vector with the hex values for a table with blue background.
}
\details{
The colour codes are based on the hex values in the Graphical profile handbook except for the body text where the cmyk value is used.
There is not agreement between the hex and cmyk values, but hex seem to give the most correct colour.

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes.R".
}
\keyword{datasets}
