% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_table_grey}
\alias{palette_table_grey}
\title{Palette with colours for table with grey background}
\format{
A vector with named colours:
\describe{
\item{black1}{Background colour for heading.}
\item{black2}{Background colour for sub-heading.}
\item{black3}{Background colour for sub-heading.}
\item{black4}{Background colour for table text.}
\item{black5}{Preferred background colour for table text.}
\item{bodytext}{Black 85\% font colour for use in body text.}
\item{heading}{White font colour for headings.}
}
}
\source{
Grafisk profil for Veterinærinstituttet versjon 2020.
}
\usage{
palette_table_grey
}
\description{
A vector with the hex values for a table with grey background.
}
\details{
The colour codes are based on the cmyk values in the Graphical profil handbook.\preformatted{The script for preparing the palette is given in \code{./data-raw/generate_NVI_palettes.R}.
}
}
\keyword{datasets}
