% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_background_per_column.R
\name{style_background_per_column}
\alias{style_background_per_column}
\title{Style cell background in Excel sheet based on column name}
\usage{
style_background_per_column(
  workbook = workbook,
  sheet = sheet,
  data = NULL,
  colnames_in_data = NULL,
  nrows_in_data = NULL,
  rule,
  colname,
  palette
)
}
\arguments{
\item{workbook}{The workbook object.}

\item{sheet}{The Excel sheet name.}

\item{data}{The data frame that have been exported to the Excel sheet. Used to
find column number and number of rows in the workbook that should be styled.
Can be \code{NULL} if both \code{colnames_in_data} and \code{nrows_in_data}
are different from \code{NULL}.}

\item{colnames_in_data}{The column names of the data frame that has been exported to
Excel, i.e. \code{colnames(data)}. Can be \code{NULL} if \code{data} has a non-NULL
argument. Values in \code{colnames_in_data} have preference for values in \code{data}.}

\item{nrows_in_data}{The number of rows of the data frame that has been exported to
Excel, i.e. \code{nrow(data)}. Can be \code{NULL} if \code{data} has a non-NULL
argument. Values in \code{nrows_in_data} have preference for values in \code{data}.}

\item{rule}{Rule that should be sourced to \code{openxlsx::conditionalFormatting}.
Defaults to "!= 0".}

\item{colname}{Vector with colour names from the palette that should be used
for styling the columns. If the column name is not the same as the colour
name in the palette, the colour name should be named with the column name
in the data frame.}

\item{palette}{Palette with named colours.}
}
\value{
None. One or more columns in the workbook object is styled.
}
\description{
Style the cell background with different colours. The background
colours are chosen based on the column names, for example "paavist" or
"tvilsom". In each column, all cells that are following the given
rule, will get the same background colour.
}
\details{
The columns will be styled with the colour in the palette that have
the same name as the column. If the column name is not the same as the
name of the colour in the palette, the vector of column names can be named
with the corresponding colour name, see example.
}
\examples{
\dontrun{
# Attach packages and set up with temporary directory
td <- tempdir()
library(NVIpretty)
library(openxlsx)

# Make workbook object
workbook <- createWorkbook()

# Add a sheet to the workbook
add_formatted_worksheet(iris,
                        workbook,
                        sheet = "iris",
                        wrapHeadlineText = TRUE,
                        collabels = TRUE,
                        colwidths = FALSE,
                        standards = NULL)

style_background_per_column(workbook = workbook,
                             sheet = sheet,
                             data = iris,
                             rule = "> 3",
                             colname = c("bluegrey" = "Sepal.Length",
                                         "bluegrey" = "Sepal.Width",
                                         "green" = "Petal.Length"),
                             palette = c("bluegrey" = "#6fb2d3", "green" = "#579e65"))

# Save the workbook
saveWorkbook(wb = workbook, file = paste0(td, "/iris.xlsx"), overwrite = TRUE)

print(paste("You may examine the Excel workbook at", normalizePath(td, winslash = "/")))
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
