% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_graph}
\alias{palette_graph}
\title{Palette with colours for graphs for 2025 - today's date}
\format{
An object of class \code{character} of length 15.
}
\source{
Grafisk profil for Veterin<U+00E6>rinstituttet version 2024 (available on
NVI's intranet).
}
\usage{
palette_graph
}
\description{
A vector with the hex values for colours to be used in graphs.
    The order of the colours are as they should be used in the graphs.
}
\details{
The colours are displayed in the vignette "NVI colour palettes".
    Remark that the colours #F2C512 and #C0ECD3 should not be used for line
    graphs.

A vector with 12 unnamed colours:
   c(#143985, #9EB8F0, #CC0E00, #FF8D85, #836A00, #F2C512, #1A5634,
    #36B46C, #C0ECD3, #5CD1FF, #737F97, #B9BFCB)

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes.R".
}
\keyword{datasets}
