% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_graph}
\alias{palette_graph}
\title{Palette with colours for graphs for 2025 - today's date}
\format{
An object of class \code{character} of length 15.
}
\source{
Grafisk profil for Veterinærinstituttet version 2024 (available on
NVI's intranet).
}
\usage{
palette_graph
}
\description{
A vector with the hex values for colours to be used in graphs. The
    order of the colours are as they should be used in the graphs.
}
\details{
The colour codes are based on the hex values in the Graphical profile
    handbook (available at NVI's intranet). The colours are displayed in the
    vignette "NVI colour palettes".

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes.R".

A vector with 15 unnamed colours:
  #59CD8B (Jord), #D7F4FF (Luft), #1C4FB9 (Vann), #091A3E (Svart),
  #BCEED1 (Lys jord), #95D9F3 (Mørk luft), #C7D9FF (Lyst vann),
  #FF5447 (Mat), #FFDAD4 (Lys mat), #F5D34A (Gul), #FBECB2 (Lys gul),
  #9A55D2 (Lilla), #DEC2F5 (Lys lilla), #FCA446 (Orange), #F9D3A9 (Lys orange).
}
\keyword{datasets}
