% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_graph}
\alias{palette_graph}
\title{Palette with NVI profile colours for 2025 - today's date}
\format{
A vector with 10 unnamed colours:
\describe{
  \item{Jord}{#59CD8B}
  \item{Luft}{#D7F4FF}
  \item{Vann}{#1C4FB9}
  \item{Hvit}{#F7FDFF}
  \item{Svart}{#091A3E}
  \item{Lys_jord}{#BCEED1}
  \item{Mørk_luft}{#95D9F3}
  \item{Lyst_vann}{#C7D9FF}
  \item{Mat}{#FF5447}
  \item{Lys_mat}{#FFDAD4}
}
}
\source{
Grafisk profil for Veterinærinstituttet version 2024 (available on
NVI's intranet).
}
\usage{
palette_graph
}
\description{
A vector with the hex values for NVI profile colours.
}
\details{
The colour codes are based on the hex values in the Graphical profile
    handbook (available at NVI's intranet). The colours are displayed in the
    vignette "NVI colour palettes".

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes.R".
}
\keyword{datasets}
