% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_NVI}
\alias{palette_NVI}
\title{Palette with NVI's profile colours for 2025 - today's date}
\format{
A vector with named colours:
\describe{
\item{Jord}{#59CD8B}{The green colour that signifies NVI's core area "Terestrial animals".}
\item{Luft}{#D7F4FF}{A light blue colour".}
\item{Vann}{#1C4FB9}{The blue colour that signifies NVI's core area "Aquatic animals".}
\item{Hvit}{#F7FDFF}{Whitish. It is used in headings with a coloured or black background.}
\item{Svart}{#091A3E}{Blackish for use in body text in reports etc. It can also be used in headings.}
\item{Lys jord}{#BCEED1}{Used in graphs.}
\item{M&#U00F8rk luft}{#95D9F3}{Used in graphs.}
\item{Lyst vann}{#C7D9FF}{Used in graphs.}
\item{Mat}{#FF5447}{The red colour that signifies NVI's core area "Food and feed safety".}
\item{Lys mat}{#FFDAD4}{Used in graphs.}
}
}
\source{
Grafisk profil for Veterin<U+00E6>rinstituttet version 2024 (available on
NVI's intranet).
}
\usage{
palette_NVI
}
\description{
A vector with the hex values for NVI's profile colours.
}
\details{
The colour codes are based on the hex values in the official colour profile.

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes.R".
}
\keyword{datasets}
