% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_NVI_profile.R
\docType{data}
\name{palette_NVI}
\alias{palette_NVI}
\title{Palette with NVI's profile colours for 2025 - today's date}
\format{
A vector with named colours:
\describe{
\item{Jord}{The green colour that signifies NVI's core area "Terestrial animals".}
\item{Luft}{A light blue colour".}
\item{Vann}{The blue colour that signifies NVI's core area "Aquatic animals".}
\item{Hvit}{Whitish. It is used in headings with a coloured or black background.}
\item{Svart}{Blackish for use in body text in reports etc. It can also be used in headings.}
\item{Lys jord}{Used in graphs.}
\item{Mørk luft}{Used in graphs.}
\item{Lyst vann}{Used in graphs.}
\item{Mat}{The red colour that signifies NVI's core area "Food and feed safety".}
\item{Lys mat}{Used in graphs.}
}
}
\source{
Grafisk profil for Veterinærinstituttet version 2024 (available on
NVI's intranet).
}
\usage{
palette_NVI
}
\description{
A vector with the hex values for NVI's profile colours.
}
\details{
The colour codes are based on the hex values in the official colour profile.

The script for preparing the palette is given in "./data-raw/generate_NVI_palettes.R".
}
\keyword{datasets}
