% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_name_vector}
\alias{set_name_vector}
\title{set_name_vector}
\usage{
set_name_vector(x)
}
\arguments{
\item{x}{[\code{character}]\cr
A unnamed, partly named or fully named vector of character variables input in add-functions}
}
\value{
A named vector where previously unnamed elements have been named with the element value as name.
}
\description{
Ensures that all elements in a vector are named.
}
\details{
Used to name all elements in vectors with column names that should
    be added to data frames. Thereby it can easily be differentiated between
    standard column names (vector elements) and wanted column names in the
    generated tables (names). The user can input a named vector to the
    function. This is considered easier for the user to understand than
    inputting two vectors.

    This function is used in different add-functions before calling add_new_variable
}
\examples{
new_columns <- c("name11" = "column1", "name2" = "column2", "column3")
new_columns <- set_name_vector(new_columns)

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
