% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_NVI.R
\name{theme_NVI}
\alias{theme_NVI}
\title{NVI theme for ggplot2}
\usage{
theme_NVI(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  legend.position = "bottom",
  ink = "black",
  paper = "white"
)
}
\arguments{
\item{base_size}{[\code{numeric(1)}]\cr
Base font size, given in pts. Defaults to 11.}

\item{base_family}{[\code{character(1)}]\cr
Base font family. Defaults to "".}

\item{base_line_size}{[\code{numeric(1)}]\cr
Base size for line elements. Defaults to `base_size` / 22.}

\item{base_rect_size}{[\code{numeric(1)}]\cr
Base size for rectangular elements. Defaults to `base_size` / 22.}

\item{ink}{[\code{character(1)}]\cr
Foreground colour. Defaults to "black".}

\item{paper}{[\code{character(1)}]\cr
Foreground colour. Defaults to "white".}

\item{legend_position}{[\code{character(1)}]\cr
Position of legend. Must be one of c("none", "left", "right", "bottom",
"top", "inside"). In official reports it should be "bottom". Defaults to
"bottom".}
}
\description{
A complete ggplot2 theme in accord with NVI's graphical profile.
    The theme control all non-data display.
}
\details{
The NVI theme without background for plot and legend, light grey horizontal support
    lines and dark grey axis' marks.
}
\examples{
library(ggplot2)
library(NVIpretty)

# Scatter plot
ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point(aes(color = Species)) +
  theme(legend.position = "right") +
  theme_NVI() +
  scale_colour_manual(values = palette_graph)

# Changing default theme for the whole session
theme_set(theme_NVI())

# The default them is now "theme_NVI" and there is no need to specify the theme within "ggplot".
# Box plot in accord with NVI's graphical profile
ggplot(iris, aes(Species, Sepal.Length)) +
  geom_boxplot(aes(fill = Species)) +
  theme(legend.position = "right") +
  scale_fill_manual(values = palette_graph)
}
